fixed a bug where fitter_routines.f95 routines would not work correctly

+++ fileio.py
@@ -1,4 +1,4 @@
 import numpy as np
 def read_file(fname):
-	time_series = np.loadtxt(fname)
+	time_series = np.loadtxt(fname, dtype=np.float64)
 	return(time_series)
